requirements.txt
torch
torchvision
matplotlib
numpy
scikit-learn
tqdm
Pillow
